requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking
#
asgiref==3.7.2
    # via django
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
crispy-bootstrap5==0.7
    # via -r requirements.in
cryptography==41.0.1
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.2.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-allauth
    #   django-anymail
    #   django-browser-reload
    #   django-crispy-forms
    #   django-extensions
    #   django-htmx
    #   django-render-block
django-allauth==0.54.0
    # via -r requirements.in
django-anymail[mailjet]==10.0
    # via -r requirements.in
django-browser-reload==1.11.0
    # via -r requirements.in
django-crispy-forms==2.0
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-environ==0.10.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-htmx==1.16.0
    # via -r requirements.in
django-render-block==0.9.2
    # via -r requirements.in
idna==3.4
    # via requests
oauthlib==3.2.2
    # via requests-oauthlib
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.7.0
    # via django-allauth
python3-openid==3.2.0
    # via django-allauth
requests==2.31.0
    # via
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sqlparse==0.4.4
    # via django
urllib3==2.0.3
    # via
    #   django-anymail
    #   requests
